INTEL_DII/FIXME: drm/i915/xehpsdv: clear the serviced bits on GEN11_INTR_IDENTITY_REG
On ICL, even if we only write GEN11_INTR_DATA_VALID all the bits will
be cleared. If we write GEN11_INTR_DATA_VALID and the serviced bits
there will be no problem either. On XEHPSDV, the documentation says that
we have to write the serviced bits, and the simulator behaves as the
documentation mandates. So switch everybody to behave like XEHPSDV needs.
Leave a big FIXME comment so we will remember to recheck it once we
have real hardware.

Acked-by: Daniele Spurio Ceraolo <daniele.ceraolospurio@intel.com>
Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>