INTEL_DII: drm/i915/xehpsdv: rework some functions to take intel_gt as argument
Interrupts in XEHPSDV are per tile: every tile has the full set of
interrupt registers. We'll have to pass around struct intel_gt
everywhere for the code to be able to know where to read registers
from.

In this commit we start passing intel_gt instead of dev_priv for the
functions that are related to XEHPSDV irq handling. Right now we're still
passing tile 0 everywhere, but in the next patches we'll start
actually passing the correct tile.

This rework is by no means complete, but it's enough for what we want
to achieve with the next patches, and it's separate so the next
patches do not get polluted with a bunch of chunks just fixing
function signatures.

Tvrtko Ursulin:

v2: Convert to struct intel_gt.
v3: with my conflict-avoidance hat: don't change gen11_irq_handler() as
    it won't be used for XEHPSDV (Lucas)
v4: __gen11_irq_handler is used by DG1 which had a conflict that was
   resolved for gen12_gt_hw_error_handler()(RK)

Signed-off-by: Radhakrishna Sripada <radhakrishna.sripada@intel.com>
Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>
Co-authored-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Signed-off-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>