INTEL_DII: drm/i915/xehpsdv: Clear CCS on object creation in LMEM
When an object is allocated on LMEM, we clear the contents of the
object. For security reasons, we should also clear the CCS associated
with the object. This patch adds a call to _i915_ctrl_surf_copy_blt() to
clear the CCS associated with the newly allocated object.

v2: Remove the clear_ccs flag. Instead, clear the CCS whenever
intel_emit_vma_fill_blt() is called with a value of 0. Change logic to
calculate the size of the batch buffer when XY_CTRL_SURF_COPY_BLT is
used.

v3: Add comments to explain when we need to clear the CCS and simplify
the logic to check if we need to clear the CCS

v4: Use correct mocs value and flush LLC and CCS after initilization
    of CCS data (Sandeep)

v5 (Lucas): Rebase on upstream

v6 (Lucas): Rebase on build breakage fix

v7 (MattR):
 - Move more logic into the common CCS copy emission function.

Signed-off-by: Neel Desai <neel.desai@intel.com>
Acked by CQ Tang <cq.tang@intel.com>
Reviewed-by: Prathap Kumar Valsan <prathap.kumar.valsan@intel.com>
Cc: Venkata Sandeep Dhanalakota <venkata.s.dhanalakota@intel.com>
Cc: Daniel Vetter <daniel.vetter@intel.com>
Cc: CQ Tang <cq.tang@intel.com>
Cc: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>
Cc: Matthew Auld <matthew.auld@intel.com>