INTEL_DII/FIXME: drm/i915/uc: fetch uc fw for each gt
The firmwares need to be loaded from the LMEM region belonging to the gt,
so we need to create a uc object for each gt.

FIXME: we could do the fetch once and copy the blob in each lmem region.
This would require some rework to decouple the version to be fetched
(global) and the object to be loaded (per-gt), which could be done in
different ways. Waiting for HW to get some performance data and decide
which way is best; in the meantime, doing the fetch 4 times is the
easiest and less invasive way to get going.

v2: do not pass intel_uc to uc_fw_fetch (Michal)

Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Cc: Michal Wajdeczko <michal.wajdeczko@intel.com>
Cc: Tvrtko Ursulin <tvrtko.ursulin@linux.intel.com>
Acked-by: Michal Wajdeczko <michal.wajdeczko@intel.com>