INTEL_DII: drm/i915/pf: Prepare runtime info for VFs
VFs can't read some of the registers but their values are necessary
to correctly use the hardware. Prepare set of the registers values
that can be shared with VF using VFPF calls.

Start with register list for TGL and ADL-P which are the same.

Unsuccessful preparation of runtime info for VFs is not critical for
the PF itself and we don't want to abort driver loading. Instead we'll
only update PF status to disallow later creation of VFs as we are
unable to meet our SLA for VFs.

v11: restored from bad squash
v12: add support for ADL-S
v13: add GEN10_MIRROR_FUSE3
v14: add GEN11_HUC_KERNEL_LOAD_INFO

Signed-off-by: Michal Wajdeczko <michal.wajdeczko@intel.com>