INTEL_DII: drm/i915/xehpsdv: Wa_1409502670 - Do not use the first 64K address space in PPGTT
In all PPGTT's, SW should Null-map the VA's at [0, 64KB - 1]--so actual
page-faults at such will be impossible.  And then driver's fault-handler
should ignore any faults in that range.

Note that as a consequence of this workaround, several selftests also
need to be updated to ensure the vma node starts after 64K.  The tests
are updated to either use either a static offset of 64k or a random
offset.

v2: Apply workaround for all A steppings (Harish)

v3 (MattR):
 - Workaround applies to all steppings now.
 - Squash in the corresponding updates we need to make to IGT self tests
   as a result of this workaround, and fix their stepping ranges too.

Signed-off-by: Harish Chegondi <harish.chegondi@intel.com>
Signed-off-by: Prathap Kumar Valsan <prathap.kumar.valsan@intel.com>
Signed-off-by: Akeem G Abodunrin <akeem.g.abodunrin@intel.com>
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>
Reviewed-by: CQ Tang <cq.tang@intel.com>