INTEL_DII: drm/i915/pvc: Enable atomics on device memory
Page table entry format for Unified Shared Memory with 57b VA support
has Atomic Enable(AE) bit in the leaf level entry. The default value of
0 doesn't grant atomic access permission to the page and any atomic
operation results in a page fault. Atomics to device memory is a valid
operation and we need to enable atomics on pages allocated in device
memory.

Signed-off-by: Prathap Kumar Valsan <prathap.kumar.valsan@intel.com>