INTEL_DII: drm/i915/xehpsdv: WA: Wa_1607720814 LNCF to l3node writes incomplete
[XEHPSDV] - LNCF not sending out updated data to l3node units

This is not a typical workaround, as the goal is not to set a register
to a particular value, but rather to perform a write operation to
within a certain address range in order to force previous writes to
that range to be completed.

In particular, with XEHPSDV, writes to address ranges 0xb000-0xb01f
or 0xb0a0-0xb0ff do not take effect until a subsequent write is
done within the same range.  Therefore, a special write is done
here to dummy register 0xb0cc to ensure that these dangling writes
are completed.  This WA is applied after the regular engine WA list
has been applied so that it will have its effect for any writes in
the list.

It has been suggested that this should be accomplished by writing 0
to 0xb0cc (which is described as a "dummy register").

Because the WA list is sorted and optimized, this special write
cannot be done just by adding it to the end of the WA list.  It is
therefore explicitly done as a separate operation after the WA list
has been applied.

v2: Apply workaround for all A steppings (Harish)

v3 (MattR): This workaround now applies to all steppings.

Signed-off-by: Harish Chegondi <harish.chegondi@intel.com>
Signed-off-by: Dale B Stimson <dale.b.stimson@intel.com>
Reviewed-by: John Harrison <John.C.Harrison@Intel.com