INTEL_DII/UAPI: drm/i915: Update VM_ADVISE for preferred location hint
We migrate BOs not on every access fault but only when required to meet
requirements for atomic access or when the faulting location is the
preferred location (as specified by user).

To introduce a preferred location, we allow the VM_ADVISE interface
to provide a preferred region. This region must be part of the placement
list set for the GEM object during GEM_CREATE.
The should_migrate function (called by fault handlers) is updated to include
test of preferred location.

Testcase: igt@gem_vm_pagefault@preferred-location
Reviewed-by: Niranjana Vishwanathapura <niranjana.vishwanathapura@intel.com>
Signed-off-by: Brian Welty <brian.welty@intel.com>