INTEL_DII: drm/i915/dg1: Remove require_force_probe protection
require_force_probe indicates that the device is recognized but
support for a platform is not completely mature. It requires either
the i915 force_probe module parameter or the
CONFIG_DRM_I915_FORCE_PROBE kernel config option.

After all required DG1 patches are in and the platform is stable,
remove the require_force_probe protection.

Acked-by: Dhinakaran Pandiyan <dhinakaran.pandiyan@intel.com>
Reviewed-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Signed-off-by: Ashutosh Dixit <ashutosh.dixit@intel.com>