INTEL_DII: drm/i915: Re-add enable_rc6 modparam
To allow easier debug of platforms which do not fully support
power-saving render C-state 6, add back the module parameter
to allow RC6 flows to be disabled. Instead of directly affecting
the RC6 states via a bitmask as done previously, use this module
parameter to clear the has_rc6 field for these platforms.

v2: Remove modparam check from rc6_supported(). Allow intel_rc6_init
to complete sanitization even when the modparam is disabled.

v3 (Lucas): Move param check to another block, after disabling rc6. It's clearer
if we have separate blocks:
	1) is it supported?;
	2) did init failed?;
	3) should we enabled it back?
This way it's easier to read that we are currently disabling rc6 even if
we are not enabling it back (when param enable_rc6=0)

Acked-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Reviewed-by: Lucas De Marchi <lucas.demarchi@intel.com>
Suggested-by: Vinay Belgaumkar <vinay.belgaumkar@intel.com>
Signed-off-by: Stuart Summers <stuart.summers@intel.com>
Signed-off-by: Vinay Belgaumkar <vinay.belgaumkar@intel.com>