INTEL_DII/FIXME: drm/i915/xehpsdv: Add Wa_14010828422
This is the potential nasty deadlock we might see on XEHPSDV A0 when there
is a lot of blitter traffic. For performance users might want to tweak
the modparam to disable the throttling to get maximum performance at
the risk of hard system deadlocks.

-add modparam "enable_hw_throttle_blt" to enble the WA.
-Configure bcs MMIO Mode register GAB 15:10 to 63 and see if we still
 hit deadlock
-Configure MERT TAGBITS 1:0 to 10 to have max outstanding RD bugger size
-Set source_target_memory = system always for FAST_COPY BLT operation

FIXME:  We should just apply this workaround conditionally and then have
a separate commit much later in the pile that makes the behavior
conditional on a modparam.  That is, assuming we even need the modparam
anymore now that PO is over...if not we should drop it.

v2 (Sandeep):
  - use correct API and location to write to register GAB_MODE
  - set MERT register values(0x2) including bit 31

v3 (Sandeep):
  We have received confirmation from HW team that MERT register should
  be programmed with TAG_10 updating the WA to use the same.
  NOTE: This WA will only stick if system is rebooted as this register
  is a write once.

v4 (Lucas - requested by Matt Roper):
  - Reword commit message
  - Make sure WA is applied only to XEHPSDV A0

v5 (Harish):
  - Apply A0 workaround for all A steppings

v6 (Lucas)
  - use wa_masked_field_set() for masked register
  - declare the mask in the register to be re-used
  - use REG_FIELD_PREP, REG_GENMASK and lowercase
  - add comment with WA number

v7 (Lucas)
  - Remove BLT_SRCMEM_SYS from intel_emit_vma_copy_blt(): xehpsdv is
    taking another branch, no need to handle step A* differently.

HSDES: 14010924592
Reviewed-by: Sudeep Dutt <sudeep.dutt@intel.com>
Signed-off-by: CQ Tang <cq.tang@intel.com>
Signed-off-by: Venkata Sandeep Dhanalakota <venkata.s.dhanalakota@intel.com>
Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>
Signed-off-by: Harish Chegondi <harish.chegondi@intel.com>