INTEL_DII/NOT_UPSTREAM: drm/i915: add modparam enable_secure_batch
gem_exec_sriov and gem_sriov_whitelisted are using deprecated flag
I915_EXEC_SECURE to validate VF malicious behavior. This is a hack to have
this tests run on platform milestones. For internal usage it was easier to
re-enable this with enable_secure_batch modparam than rewritting tests as selftests.
This is the only reason to have this patch.

v2: change modparam param from int to bool, add HACK to patchname.
v3: Update patch to work with multi-bb execbuf changes
v4 (Lucas): Rebase on upstream and misc changes

Cc: Piotr Piórkowski <piotr.piorkowski@intel.com>
Cc: Katarzyna Dec <katarzyna.dec@intel.com>
Cc: Antonio Argenziano <antonio.argenziano@intel.com>
Signed-off-by: Daniel Mrzyglod <daniel.t.mrzyglod@intel.com>
Acked-by: Piotr Piórkowski <piotr.piorkowski@intel.com>