INTEL_DII: drm/i915/iaf: Introduce a module parameter to control IAF
Add a new module parameter to control enable/disable of the IAF
feature.

This module parameter is needed because SRIOV and IAF are mutually
exclusive (see HSDES #14014623804).  Until this is fixed, (planned for
the C0 stepping) the driver needs to be able to enable/disable the IAF
infrastructure. And at the moment this is the only way to do this.

v1: default true (mjruhl)
v2: add SRIOV text
v3: since this is going to be a permanent option, remove _unsafe

Reviewed-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Acked-by: Sudeep Dutt <sudeep.dutt@intel.com>
Signed-off-by: Michael J. Ruhl <michael.j.ruhl@intel.com>