INTEL_DII/NOT_UPSTREAM: drm/i915/xehpsdv: Implement Wa_1409759559
If we enable RC6 on a multi-tile A-step platform, we need to use
DFD_RESTORE to re-program the XEHPSDV_TILE registers when the GT wakes from
RC6 in multi-tile setups.  This DFD_RESTORE mechanism isn't functional
on A0 stepping, but can be applied on A1+.

Note that RC6 should already be disabled on all multi-tile A-step due to
other workarounds (e.g., Wa_16011524941) so we generally can't end up in
a situation where we need to apply this DFD_RESTORE workaround.  However
silicon validation still plans to run RC6 experiments on these steppings
and may overrule that stepping check manually with the
rc6_ignore_steppings module parameter, so we still need to implement
this additional workaround to prevent them from falling on their faces
when they do so.

v2 (MattR):
 - Fix incorrect bit splitting of the physical address between the two
   DFD_RESTORE registers.  Also add a clarifying comment about why we
   don't want to program the 'lock' bit of the register.

v3 (MattR):
 - Update commit message to clarify that this is now only useful for SV
   experiments.  Move this patch later in the XEHPSDV pile (after the
   rc6_ignore_steppings parameter is added) and mark it NOT_UPSTREAM.
 - Drop the A0 test in xehpsdv_rc6_init; if SV gets here on an A0, then they
   hopefully know what they're getting themselves into.

v4: Rebase

v5 (MattR):
 - Drop XEHPSDV_ prefix; the registers have been at this offset since
   TGL but we just haven't needed to use them until now.

Cc: Matt Roper <matthew.d.roper@intel.com>
Cc: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Cc: John Harrison <john.c.harrison@intel.com>
Signed-off-by: Harish Chegondi <harish.chegondi@intel.com>
Signed-off-by: Caz Yokoyama <caz.yokoyama@intel.com>