INTEL_DII: drm/i915/pmu/xehpsdv: Export counters from all tiles
Start exporting frequency and RC6 counters from all tiles.

Existing counters keep their names and config values and new one use the
namespace added in the previous patch, with the "-gtN" added to their
names.

Interrupts counter is an odd one off. Because it is the global device
counters (not only GT) we choose not to add per tile versions for now.

v2: validate gt_id to be with in the tile count

v3:
 * Always add -gtN suffix on multi-tile systems.

v4:
 * Do not append gt suffix on global (not per tile) events.

Signed-off-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Acked-by: Andi Shyti <andi.shyti@intel.com>
Signed-off-by: Aravind Iddamsetty <aravind.iddamsetty@intel.com>
Reviewed-by: Michael J. Ruhl <michael.j.ruhl@intel.com> # v2
Cc: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>