INTEL_DII/FIXME: drm/i915/vf: Don't try to update GuC policy if VF
In commit ("drm/i915/guc: Connect reset modparam updates to GuC
policy flags") we started notifying GuC about changes in scheduling
policies, but such notification is not applicable to VF.

FIXME: GuC spec: "Note: This is not a production usage flow."
FIXME: pending ABI changes for execution and preemption time quantums

v2: Add FIXME comment to code. Also, note that 'not applicable' above
should be 'not permitted' and a failure code needs to be returned to
userland (JohnH).

Signed-off-by: Michal Wajdeczko <michal.wajdeczko@intel.com>
Acked-by: Michał Winiarski <michal.winiarski@intel.com>