INTEL_DII/UAPI: vm_bind: Add User Fence support
Currently, we only allow waiting on the forward progress of an individual
GEM object, or of a GEM execbuf fence. The primary purpose of the fence
is to provide a scheduling primitive to order the execution flow of
batches (cf VkSempahore).

Userspace instead uses values in memory to implement client fences, and
has to mix busywaiting on the value coupled with a dma_fence in case it
needs to sleep. It has no intermediate step where it can wait on the
memory value itself to change, which is required for scenarios where the
dma_fence may incur too much execution latency.

The CPU equivalent is a futex-syscall used to setup a waiter/waker based
on a memory location. This is used to implement an efficient sleep for
pthread_mutex_t, where the fast uncontended path can be handled entirely
in userspace.

This patch implements a similar idea, where we take a user process
virtual address and install an interrupt handler to wake up the
current task when the memory location passes the user supplied filter.
We term this as user fence.

It also allows the user to emit their own MI_FLUSH/PIPE_CONTROL notify
interrupt  within their batches after updating the value on the GPU to
have sub-batch precision on the wakeup.

User fence <user address and value pair> can also be supplied to the
kernel driver to signal/wake up the user process after completion of an
asynchronous operation.

This patch is derived from the below orignial work:
https://patchwork.freedesktop.org/patch/349417/

Opens:

- on attaching the waiter, we enable interrupts on all engines of the
context irrespective of which are active.
 * we can optimise when to enable interrupts while the context is active
 * we can extend the interface for the user to select which engines may
   wake us

- we could return an fd wrapping the comparison operation on the memory
address if we want to pass the waiter around different processes or
reuse the waiter (with poll() + read() like timerfd).

- Handle engine reset while user waiting on user fence

References:
- b2c97bc78919 ("anv/query: Busy-wait for available query entries")
- https://gitlab.freedesktop.org/mesa/mesa/merge_requests/3279#note_377240
Testcase: igt/gem_wait_user_fence

v2: Rebased (Niranjana)
v3: Remove worker thread (Niranjana)
v4: Fix address range check (Niranjana)
v5: Handle timeout properly (Niranjana)
v6: Downstream PRELIM uapi changes (Ashutosh)

Signed-off-by: Chris P Wilson <chris.p.wilson@intel.com>
Signed-off-by: Niranjana Vishwanathapura <niranjana.vishwanathapura@intel.com>
Reviewed-by: Prathap Kumar Valsan <prathap.kumar.valsan@intel.com>