INTEL_DII: drm/i915: Check eu attentions on heartbeat
If there is thread attention set, it means we can't make
progress until it is resolved. For now we can't resolve it
if there is no debug/sip handler for it so issue error handler
to resolve the situation.

v2: dont gate with modparm, remove check in parks (Chris)
v3: check only if intel_engine_has_eu_attention (Dominik)

Signed-off-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>
Reviewed-by: Maciej Patelczyk <maciej.patelczyk@linux.intel.com>