INTEL_DII: drm/i915/vf: Ignore GuC CT event from memIRQ if CT is disabled
Platforms supporting more than 7 VFs, use memory based interrupts to
signal events from GuC.
In case, when we are disabling CT on VF, we are unable to mask memory based
interrupt from GuC, so there is a chance that an GuC CT event for VF will come
just as CT will be already disabled. As we are unable to properly handle such
an event, so we must abandon it.
Let's start ignoring this event without any consequences when CT is
disabled and event come form memory based interrupt.

Jira: VLK-12945

Signed-off-by: Piotr Piórkowski <piotr.piorkowski@intel.com>
Cc: Michal Wajdeczko <michal.wajdeczko@intel.com>
Cc: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Reviewed-by: Lucas De Marchi <lucas.demarchi@intel.com>