INTEL_DII/NOT_UPSTREAM: drm/i915/xehpsdv: Hardcode the hwconfig for XEHPSDV
XEHPSDV does not support hardware config tables in the hardware. Instead,
the table must be included in the i915 driver or not at all. Note that
the POR is for 'not at all'. This patch is to be kept only until the
UMDs have been updated to not require the table.

Signed-off-by: John Harrison <John.C.Harrison@Intel.com>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Reviewed-by: Matthew Brost <matthew.brost@intel.com>