INTEL_DII: drm/i915: Add the reset count to intel_engine_cs
When referencing the reset_engine_count with only class we fail to access
corresponding engine entry rather we limit upto no of class entries only
for any engine. So if a particular engine was reset we end up showing
reset for all the engines belonging to that class. If we maintain
the counter in intel_engine_cs we can capture resets for all instances

Refer: cb823ed9915b ("drm/i915/gt: Use intel_gt as the primary object for handling resets")

Cc: Chris Wilson <chris.p.wilson@intel.com>
Cc: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Cc: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>
Signed-off-by: Aravind Iddamsetty <aravind.iddamsetty@intel.com>
Reviewed-by: Chris Wilson <chris.p.wilson@intel.com>