INTEL_DII: drm/i915/guc: Count no of engine resets at GT level
In GuC submission mode identifying which individual engine was reset is
not  currently possible. All the i915 driver is actually told is the
context that was reset. There will be a GuC instance per GT, so it should
be enough to tally up all the errors from all contexts of a GuC into a
number and tag it to that GT.

This counter is also incremented in execlist mode for engine resets
associated with a GT.

v2:
1. in GuC submission mode report engine resets at GT level
2. return -1 for the reset counts on GuC submission (Tvrtko)
3. on error capture we should also have "engine_reset_count_total" or
so to have an understanding if the machine has been resetting like crazy
before it hanged (Joonas)

v3:
1. Count the engine resets at GT level for both GuC and execlist
submissions (Tvrtko)
2. maintain the counter as top level gt structure as it can be kept
regardless of the backend(Tvrtko)

v4:
1. rename the print in capture log to GT total engines reset count

v5:
1. move engines_reset_count inside intel_reset(Chris)

Cc: Chris Wilson <chris.p.wilson@intel.com>
Cc: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Cc: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>
Cc: John Harrison <John.C.Harrison@Intel.com>
Signed-off-by: Aravind Iddamsetty <aravind.iddamsetty@intel.com>
Reviewed-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>