INTEL_DII/UAPI/NEEDSIGT: drm: Add Client Cap for advance gamma mode
Introduced a client cap for advance cap mode capability. Userspace
should set this to get to be able to use the new gamma_mode property.

If this is not set, driver will work in legacy mode.

Note: This is suggested by Ville and based on his idea, the new
gamma mode handling is designed.

Signed-off-by: Uma Shankar <uma.shankar@intel.com>
Reviewed-by: Bhanuprakash Modem <bhanuprakash.modem@intel.com>