INTEL_DII: i915/perf: Whitelist upper word of EXECLIST_STATUS reg
Currently user issues an MI_REPORT_PERF_COUNT to obtain the hw context
id used in OA reports. This command would cause OAR unit to write an OA
report to the user buffer and user would read the context ID field of
the OA report to get the current context ID. The command requires
setting up the OAR unit in addition to the OA buffer used to capture the
periodic reports. In addition, the definition of the context ID field in
the OA report is not guaranteed to have the same semantics for
different types of reports and platforms.

Provide a simpler and consistent mechanism for a user batch to read its
own context ID by whitelisting the upper dword (context id) of the
EXECLIST_STATUS register. The register itself is context saved/restored.

Signed-off-by: Umesh Nerlige Ramappa <umesh.nerlige.ramappa@intel.com>
Reviewed-by: Ashutosh Dixit <ashutosh.dixit@intel.com>