INTEL_DII/NOT_UPSTREAM: drm/i915/gt: page allocation for HW debug
The GT HW supports several types of workaround BB, but we currently do
not support all of them. In HW debug scenarios (especially rc6 ones),
the SV team often wants to add and/or modify one of those BBs. Long
term we'll need an infrastructure to do those modification via the
driver, but short term the SV team has agreed to do most of the
programming via their tools. They still require the driver to do the
memory allocation and ggtt programming to avoid conflicts, so this patch
adds a mechanism to allocate extra pages during driver load. A debugfs
has also been added to dump the offset and content of the pages.

Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>