drm/i915/gt: Reduce GT runtime stats from seqlock to a latch
Since we can compute the elapsed time to add to the total, during the
PMU sample we only need to have a consistent view of the (start, total,
active) tuple to be able to locally determine the runtime. That can be
arrange by a pair of memory bariiers and carefully sequencing of the
writes and reads.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Andi Shyti <andi.shyti@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20210130160041.16897-1-chris@chris-wilson.co.uk
(cherry picked from commit (dropped, was: 963ef479f089e90f09b2231d0024c6491e32d9cc))