drm/i915/gt: Reduce engine runtime stats from seqlock to a latch
Since we can compute the elapsed time to add to the total, during the
PMU sample we only need to have a consistent view of the (start, total,
active) tuple to be able to locally determine the runtime. That can be
arrange by a pair of memory bariiers and carefully sequencing of the
writes and reads.

v2: Rebase from GuC engine busyness support (Umesh)

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Andi Shyti <andi.shyti@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20210130160041.16897-2-chris@chris-wilson.co.uk
(cherry picked from commit (dropped, was: 94a42420c79c6ccc32c045d84fe1608468ef8e70))