INTEL_DII: drm/i915: RC6 related WAs for PVC
This includes following WAs:
- 22011383443
- 22012237902
- 1508652630
- 1509372804

v2: WA is for BD, not CT
v3: Disable coarse PG instead of just returning
v4: Review comments (Matt Roper)
v5: Squash the remaining two RC6 WAs into this one.
The only PVC SKU that has RC6 enabled now is XL 1T.
For 1509372804, this is a partial WA. the remaining
part includes adding a set of save/restore registers
either in GuC or KMD. Please see JIRA-25153 for details.

v6(Tilak):  Add Wa_1509372804
At RC6 exit, there is an inadvertent error injection happening,
This is causing corrupted parity reaching the L3Bank leading to
GT fatal error.
Adding i915 opt-in for second part of WA,to enable WA in render
force wake step by GuC FW,before any work submission to CCS engines.
This is FIXME, as part of the WA, adding a set of save/restore
registers either in GuC or KMD. is not implemented and is still
in discussion with GUC FW and Hw team (VLK-25153).(Rodrigo)

v7: Squash to RC6 related WAs commit( Lucas)

v8(Tilak):

Wa_1509372804 Fix Stepping bound check,The i915 opt-in is
applicable for CT XT B0, so corrected the stepping bound.(Matt)

v9: Work is in progress B0 BD 2T HW issues 16015476723 and 16015496043.
This is a temporary fix to disable RC6 on all B0 until the software Wa
is finalized and tested. There are issues with 1T RC6 as well where
pcode WAs are being tried out - 16015350383 & 16015295493, disable RC6
on 1T until the fixes are merged.

v10: Rebase (Umesh)

v11(Tilak):
Guard i915 opt-in for Wa_1509372804 with enable_rc6,
the WA is getting enabled irrespective of enable_rc6,
so fixed it.

v12: Recommendation to disable RC6 on all PVC B-step (hsd:14015706335)

Cc: Matt Roper <matthew.d.roper@intel.com>
Signed-off-by: Vinay Belgaumkar <vinay.belgaumkar@intel.com>
Signed-off-by: Tilak Tangudu <tilak.tangudu@intel.com>