drm/i915/perf: Flush context writes before destroying the perf stream
As the context state manipulation is running independently of the perf
stream destruction, it is possible for the i915_oa_init_reg_state() to
continue to access the perf stream after it is freed. Reintroduction
serialisation between i915_oa_stream_destroy and i915_oa_init_reg_state
by virtue of rcu read locks and synchronisation.

Fixes: feed5c7be22c ("drm/i915: Pin the context as we work on it")
Closes: https://gitlab.freedesktop.org/drm/intel/-/issues/3156
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: <stable@vger.kernel.org> # v5.6+