INTEL_DII: drm/i915/selftest: Don't use the 'bind' engine for selftests
Some of the hangcheck selftests fail if run on the L4 WA's 'bind' engine
(BCS0). This is because they block/reset that engine while at the same
time the driver requires bind operations to complete for the test to
make forward progress. So, make sure those tests don't run on that engine.

v2: Nobble the hangcheck 'active others' tests as well (JohnH)

Signed-off-by: John Harrison <John.C.Harrison@Intel.com>
Reviewed-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>